Added a millisecond timer to the threading library.
Since we are removing boost threads shortly, and since the boost::date_time has issues with thread safety on osx, we moved the code in CTimeUtils::GetTimeMS() into XbmcThread::currentClockMillis(). It was also adjusted so that the windows call to timeGetTime handles the wrapping correctly (which it didn't appear to do before)."
