+++ xbmc/threads/Time.h
@@ -1,0 +1,35 @@
+/*
+ *      Copyright (C) 2005-2011 Team XBMC
+ *      http://www.xbmc.org
+ *
+ *  This Program is free software; you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation; either version 2, or (at your option)
+ *  any later version.
+ *
+ *  This Program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with XBMC; see the file COPYING.  If not, write to
+ *  the Free Software Foundation, 675 Mass Ave, Cambridge, MA 02139, USA.
+ *  http://www.gnu.org/copyleft/gpl.html
+ *
+ */
+
+#pragma once
+
+namespace XbmcThreads
+{
+  /**
+   * This function returns the system clock's number of milliseconds but with
+   *  an arbitrary reference point. It handles the wrapping of any underlying
+   *  system clock by setting a starting point at the first call. It should
+   *  only be used for measuring time durations.
+   *
+   * Of course, on windows it just calls timeGetTime, so you're on your own.
+   */
+  unsigned int currentClockMillis();
+}
